backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
load-dotenv==0.1.0
Markdown==3.7
pika==1.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
whitenoise==6.9.0
